+++ .env.test
@@ -21,7 +21,7 @@ STORAGE_SECRET=8cad66340bc92edbae2ae3a792d351f48c61d1d8efe7b2d9408b0025c1f7f845
 TUNNEL_PORT=8080
 WORKER_PORT=8081
 TUNNEL_SUBDOMAIN=test-github-app-instance
-PRIVATE_KEY_PATH=./test/setup/test-key.pem
+PRIVATE_KEY_PATH=./src/common/test-utils/setup/test-key.pem
 ATLASSIAN_URL=https://test-atlassian-instance.net
 TRACKING_DISABLED=true
 HYDRO_BASE_URL=https://hydro-base-url.com/api/v1/events
