+++ src/backend/jira/enable/enable.test.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import testTracking from "../../../../test/setup/tracking";
+import testTracking from "../../../common/test-utils/setup/tracking";
 import enable from ".";
 import { Installation } from "../../models";
 import { mocked } from "ts-jest/utils";
