--- test/setup/test-key.pem
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpAIBAAKCAQEAqiSQ52bOYo9R/OZ9RzfJO2TZ05m9d+KtBhn5e+AwgJsSWFQK
-gjPARe4Kj5mX69ltes9lAB4Dl2I6Yj8SF4JM44qk1NmeGigsIKzDF+T02dov4Xbe
-/lgqod0qlBJHwecBcCVGu27490O0ccGDjhzZFuTqDvU9+h0hJrcN1pdNxmCOCVMU
-EkfaPNnBoIxgCUbh0+R4HMmT9gQJfkfpFCwKE3AqBuWQH1efxM8r/QwX/IRSMdSb
-/9Z7U9w2BGgTaEjJthWY3lROtvKpLgofNEZ6+l8Hvnwfg+1k9znZMgpTb37DCyV0
-YeDhbP9uM5rw/kAErtmh3BA4dNERnztv1EtnsQIDAQABAoIBAGX74wuBWQZE6B0S
-P1eX/bvnOJYYWjRpH9M/cIarHAeKbBWb0my2/nIsPBJVO3X4byIk/GJ0myx7iYlo
-SrzispnVefXHt3fwSnsPcyV4bLfEi9y2dvW3/V3mhuYYiwCUKg2BmoMxAnScFaX2
-G5fsfMe3UQDESXEJUty07suJB+p27OjtpkrKbyXtQo5l6fGVvkXTwwvD0H6Ujlir
-lU2VZbrQ+kuKbzU7gEgyxb63reSbz6wzXxZtvNAGd2eJ6Tn7UoIufzNN5zJvNvEJ
-7m8cOeGjQZKCybXr7Jij0VWQOdW/NQzsoLEzbkXC6LL5sEiz/olZRpAmXfzGL9Ao
-liozCC0CgYEA4NA+EhOpZMg/j1gUMUhRTId1qpfRjMuzaM6jMnSfI1NANA5MuTQb
-Bn6ckYOTmuvgtW7doBB5St1VP2Ydfkl8gZE6ce7Wm+F2CyAuMtTKY5PcrgsV9uxR
-8phkL8xcuP2C2BtayA8YUifeTwJ310ZYEfFhHdHXPq7y2aNBfBaBwBcCgYEAwb7R
-BDUbm6ya7FPK6U6tDU75YjkgiOmYnOX7Jw6G9LwDENYmok3nNkrTAjZj7IvXZk6X
-/+sIDwAuFzwlZH/CERfZNSS553YmTmdrpF5SCC2T0GfzmRmyv1zo0kFPh86tzvqL
-6YYdloP2nBML5HQzeFHNGscPjCIUb0v3uKRK63cCgYBlKC6V0bUJ/TTkWhyT5Fd4
-mQcwIKoQsWPkB/7U6DbnHUGparnJKBH1BSmMLGlRb0uoN9+Pc3owXpFKOreQqJ5W
-arOiqnZ0o6mK7gr7olYDInhM8/4JImLpRvrlqw4ZyzCB7yUVGf22kevtH04vVp6w
-TC2HznEM6omptlTz8eimNwKBgQCK3jh2qZI2TMtmZNDl1SyYg2igjI3TVIu8ueMA
-pZr5w+D2qmhaoTf+ZDdGOLH6lGk5vU+ZcV7/UYO3x7aNiD0Tf2KCguZ1Dux5+grf
-b3OgxPpTwI6LyC2kl/WQ7psQnoVZ+59ZLHX8sm2KhORbRSa3MPqwHMn6CVtYQr02
-LRw+FQKBgQCO6UUzu+/moxIXrMYNF1NRYES/zW6l8oz+eMJN1mhhLmSdUnoc7e1u
-7UMm76ASBLMAG9HgmqSQd/eg31NRHJSGPzm8s2kBmiFb5e/HPyyabjnRq2gM8Ca3
-uNTbIr5pMXeM0ofkWmjzdXhV5dJWwnUTq2cPUvwaRjRy5iWDUkMJAA==
------END RSA PRIVATE KEY-----
